Fix for minor typo
Hi - Working through this tutorial. I believe line 436 should read

- `covariateId` identifies the covariate, and definitions of covariates can be found in the `covariateData$covariateRef` object

not

- `covariateId` identifies the covariate, and definitions of covariates can be found in the `cohortData$covariateRef` object. 

I can't find reference to a cohortData object in the vignette and `covariateData$covariateRef` appears to hold the referenced information. 

Apologies if this is wrong (and if so, please let me know where I Can find the cohortData object).